AArch64: refactor code to support set GIC device entity
In AArch64 systems, the state of GIC device can only be
retrieved from `KVM_GET_DEVICE_ATTR` ioctl. Therefore to implement
saving/restoring the GIC states, we need to make sure that the
GIC object (either the file descriptor or the device itself) can
be extracted after the VM is started.

This commit refactors the code of GIC creation, which ensures the
GIC object is not lost in `arch::configure_system` after creation.

Change-Id: Ia04c1fbc64914d85707939ec0a25090a7e3e06c9
Signed-off-by: Henry Wang <Henry.Wang@arm.com>
Jira: ENTOS-1753